requirements.txt
python-dotenv
ccxt
aiohttp
solders
base58